Remove OAuth client functionality.
PA's SPA features allow a HTTP-only cookie to be used for all API
access. As a result, the OAuth client functionality can be removed from
the SPA in favor of a simple amount of code that establishes the PA
session after the user logs in.

This also enables the OAuth client functionality to be removed from
the SPA.